+++ src/main/java/presentation/EditTweetMenu.java
@@ -8,6 +8,7 @@ import java.util.Scanner;
 
 public class EditTweetMenu implements Menu {
     Scanner console = new Scanner(System.in);
+
     @Override
     public Menu action() {
 
@@ -18,27 +19,34 @@ public class EditTweetMenu implements Menu {
             Long id = console.nextLong();
             if (id == -1) return new InsideMenu();
 
-            List<Tweet> tweets = ApplicationContext.getInstance.getUser().getTweets();
-            for (Tweet tweet : tweets) {
-                if (tweet.getId() == id) {
-                    myTweet = tweet;
-                    check = true;
-                    System.out.printf("%s %n %s %n ","1-rewrite the tweet","2- delete tweet ");
-                    int choice = console.nextInt();
-                    switch (choice){
-                        case 1: tweet.setTweet(editTweet());
-                        ApplicationContext.getInstance.getTweetService().update(tweet);
-                        ApplicationContext.getInstance.setTweets(ApplicationContext.getInstance.getTweetService().showAllTweets());
+            List<Tweet> tweets = ApplicationContext.getInstance
+                    .getTweetService().findUserTweets(ApplicationContext.getInstance.getUser());
+            if (tweets.isEmpty()) System.out.println("you have not any tweet yet");
+            else
+                for (Tweet tweet : tweets) {
+                    if (tweet.getId() == id) {
+                        myTweet = tweet;
+                        check = true;
+                        System.out.printf("%s %n %s %n ", "1-rewrite the tweet", "2- delete tweet ");
+                        int choice = console.nextInt();
+                        switch (choice) {
+                            case 1:
+                                tweet.setTweet(editTweet());
+                                ApplicationContext.getInstance.getTweetService().update(tweet);
+                                ApplicationContext.getInstance.setTweets(ApplicationContext.getInstance.getTweetService().showAllTweets());
+                                break;
+                            case 2:
+                                ApplicationContext.getInstance.getTweets().remove(tweet);
+                                ApplicationContext.getInstance.getTweetService().delete(tweet.getId());
+                                tweets.remove(tweet);
+                                break;
+                        }
 
-                        case 2: ApplicationContext.getInstance.getTweets().remove(tweet);
-                        ApplicationContext.getInstance.getTweetService().delete(tweet.getId());
-                        tweets.remove(tweet);
                     }
 
+                    if (tweets.isEmpty()) break;
+
                 }
-
-
-            }
             if (!check) System.out.println("you have not any tweet with this id try again or for exit enter -1");
 
         } while (!check);
@@ -48,9 +56,10 @@ public class EditTweetMenu implements Menu {
 
     }
 
-    public String editTweet(){
+    public String editTweet() {
+        Scanner scanner = new Scanner(System.in);
         System.out.println("write a new tweet");
-        return console.nextLine();
-
+        String edited = scanner.nextLine();
+        return edited;
     }
 }
